Add SpaceBeforeScript linter
Add linter that checks for a space following the '-' or '=' symbols.

An unexpected complication arises when dealing with tags with inline
scripts that span multiple lines via a trailing comma or explicit
multiline vertical bars. In this case, we need to extract the text from
all lines and compare it against the parsed inline text (which is
stripped of leading/trailing whitespace so we can't check it easily) in
order to find out if a separating space exists.

The specs should describe these cases adequately.

In anticipation of many specs of this nature, add a shared example
context which normalizes any `haml` variable defined via `let` so that
the specs can be significantly DRYed up.

Change-Id: I286d6956e79cbbf73b5cc062765c78bf242a063e
Reviewed-on: https://gerrit.causes.com/28446
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>