Add ScriptExtractor visitor
Add a visitor implementation which can extract Ruby code from a HAML
parse tree, storing both the extracted source code and a mapping of the
source code lines to the lines in the original HAML file.

This source code will be passed to a Ruby linter, which can then be used
to find lints in the original HAML via the source map.

Change-Id: Ifbb569f1a86185b9ce2a07006c6e157c63597b79
Reviewed-on: https://gerrit.causes.com/28448
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>