Fix ScriptExtractor to work with for/while loops
We weren't properly generating code for `for` and `while` loops in HAML
documents. This was due to the fact that the regex we were borrowing
from the HAML parser did not account for these types of statements.

Modify our block keyword detection to include the two keywords `for` and
`while`.

Change-Id: I928ef3975842232e0f5a7cc151f698d18cd27eff
Reviewed-on: https://gerrit.causes.com/28451
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>