Relax HAML dependency
The HAML dependency was very specific, which resulted in problems when
this gem was included alongside gems such as `html2haml`.

While the intent of pinning at this specific version was originally to
mitigate problems due to potential monkey patching, this didn't end up
being a requirement.

Relax the dependency to require HAML 4.0+.

Change-Id: Ie5281f2eb9f52fb65ae52d229823238865fc83ff
Reviewed-on: https://gerrit.causes.com/33567
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>