Fix loading correct .rubocop.yml config
The problem was caused by the fact that haml-lint creates temporary Ruby
files which is passed to Rubocop. The temporary files were outside of
the repository (in the `/tmp` directory), so Rubocop doesn't pick up the
corresponding `.rubocop.yml`.

Fix by creating the temporary Ruby file in the same directory as the
HAML file.

Change-Id: Ib3b2f3075a0c54ca565134d4c660337f39b31991
Reviewed-on: http://gerrit.causes.com/36432
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>