Skip visiting EOD HAML comment
For some reason that isn't terribly clear from the parser's code, the
HAML parser appends an empty HAML comment to the end of all HAML
documents.

Since we'll have linters that conceivably check comments, we don't want
them to have to be aware of this "special" comment. Add some logic to
the `HamlVisitor` so it is ignored.

Change-Id: Ia151dbbe6f4c9d1247164f2a6b3c14643160d48d
Reviewed-on: http://gerrit.causes.com/42348
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>