Fix bug in UnnecessaryInterpolation
The UnnecessaryInterpolation linter would incorrectly report a lint for
tags with content that contained interpolation at the beginning, e.g.

  %tag #{interpolation} -- #{more_interpolation}

This was due to the really naive way we were checking if a string only
contained interpolation (only check if the first two characters were
`#{`).

Solve this by using the `extract_interpolated_values` helper originally
defined in the `ScriptExtractor`, but is now pulled out into the
`HamlLint::Utils` module.

Change-Id: I6eb1b2c06b4df0a3f6dc896e089636c853464bb1
Reviewed-on: http://gerrit.causes.com/42566
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>