Support multiline comments for `Node#text`
Fetching the text associated with a HAML comment was not returned past
the first line for multiline comments. Fix this by override the `text`
method to fetch each subsequent line.

This required the introduction of a `Node#successor` method to serve the
same purpose as `Linter#next_node`, but on the `Node` itself.

Change-Id: If97def771749185c9a67fddb1afcc9a244747058
Reviewed-on: http://gerrit.causes.com/44991
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>