Improve implementation of source_from_first_line
The name of this method was a bit misleading and its implementation
would return all the source code of the document starting from the line
the node resided on.

To improve the situation, change the implementation to return only the
source of the node itself and add tests to ensure correct behavior.

Change-Id: I2372ac4c5333c5fa84c72100f5a2bf66f85d7182
Reviewed-on: http://gerrit.causes.com/45301
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>