Don't output plain text nodes in comments
Previously, the `ScriptExtractor` would output the content of a plain
text node as a `puts` statement followed by a comment containing the
actual text.

This had an unintended side-effect of setting off the RuboCop
`Style/AsciiComments` cop for text with Unicode characters.

Since this text is only for debugging purposes and isn't actually
important to the semantics of the generated code, we can safely omit it
to avoid this problem.

Change-Id: I1bd22dfbe7e443ac5d3a3dd2e620c49eb1fdfa72
Reviewed-on: http://gerrit.causes.com/45398
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>