Improve HamlLint::RakeTask
The tests for this task were sparse and not testing everything they
should have been. In the process, I noticed that the way this had been
implemented didn't entirely match up with the practices carried out in
the `rake` task libraries, so I updated the approach to be more in line
with them.

These modifications also allow us to now specify an explicit list of
files via the command line.

Sneaking in here is a change that also allows `haml-lint` to
automatically handle file glob patterns passed via command line
arguments (i.e. ones which have not already been expanded by a shell).

Change-Id: I6cb7772eeef0ee0cf8bf94cea2f0f7b8485940ed
Reviewed-on: http://gerrit.causes.com/45642
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>