Remove color-deciding concern from Logger
Now that we provide the `--[no-]color` flag, it feels wrong to have the
`Logger` class be responsible for deciding whether or not output is
colored.

Instead, we extract that decision out into the `CLI` class, who decides
it based on the `color` option, defaulting to whether or not the logger
is outputting to a TTY if it is not specified. This changes the
behavior slightly from the original implementation, as that version
would always pick based on whether it was outputting to a TTY, only
consulting the color option if it wasn't a TTY.

Change-Id: Iad332ea21f78e6853874632c2a41dcc6b848188e
Reviewed-on: http://gerrit.causes.com/45649
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>