Incorporate slim-lint design back into haml-lint
The `slim-lint` project was created using `haml-lint` as a template, but
then improved on it to make 100% code coverage possible.

Incorporate the changes back into `haml-lint`, since much of the
framework code is identical. This paves the way for us to consider
extracting the framework code into a separate gem.

This includes a bunch of specs so we can get `haml-lint` closer to 100%
code coverage.