Switch RuboCop mode to read from stdin
Currently, haml-lint runs the RuboCop linter from a Tempfile. This works
well in traditional environments, but fails in constrained environments
like Docker containers that have their data mounted as read-only
volumes.

This changes the RuboCop linter to run via an overridden STDIN. By
overriding the global `$stdin` variable, we can trick RuboCop into
believing it is reading directly from the STDIN. This means that we
don't have to generate Tempfiles for each file we want to lint which
will allow haml-lint to work within the confines of a Docker container.