Report syntax errors as a "Syntax" linter
Currently, syntax errors are reported without a linter. This is
inconsistent with the way the rest of the lint is reported, so it makes
sense to wrap syntax errors in a similar construct as the rest of the
lint.

Since syntax errors are detected via a `Haml::Parser` that bubbles up an
exception to the `Runner`, the only place to currently put the logic is
within the `Runner` itself. This makes the `HamlLint::Linter::Syntax`
class look funny because it doesn't lint on a Haml AST - it lints on the
absence of an AST.