Add support for Haml 5
Haml 5 is has been in Beta 2 in their repository for a while and has
some breaking API changed when compared with Haml 4. By adding an
adapter layer, we can maintain compatibility with both Haml 4 and 5.

To test on both Haml 4 and Haml 5, this introduces the Appraisal gem
that allows you specify different Gemfiles to test against. There was
a slight incompatibility with the Overcommit tool that we use during our
build process that required a tweak to the Overcommit configuration.

Overcommit's bundler integration works by overriding the
`BUNDLER_GEMFILE` environment variable. Appraisal works by specifying
this environment variable before running a command. The two were
conflicting and Overcommit was always overriding the Gemfile.

By disabling this feature, we can test on both Haml 4 and Haml 5 with an
easy-to-use command, yet still use the Overcommit gem as intended, since
it is present in both of the generated Gemfiles.

Closes #152