Add a linter for checking id name formats
Haml style should be consistent in the naming of things. In particular,
we should be consistent in the format that our tag ids have. This linter
lets you customize the format of your ids to one of four formats:

* `lisp-case` like the Bootstrap convention (default)
* `camelCase` like the JavaScript variable convention
* `PascalCase` like the Ruby class name or React component convention
* `snake_case` like the Ruby variable convention

This is configured in the `haml-lint.yml` file as follows:

```yaml
linters:
  IdNames:
    enabled: true
    style: <STYLE>
```

where `<STYLE>` is one of: `lisp_case`, `camel_case`, `pascal_case`, or
`snake_case`.

Closes #138