Ensure we always have a linter name
If an anonymous class inherits from `HamlLint::Linter`, the `#name`
method would fail when called because the anonymous class has a `nil`
name.

This change ensures we at least get a string for the name, which lets
specs pass when using the `#name` attribute on a test anonymous Linter
class.