Ensure RepeatedId linter resets between files
When I originally wrote the RepeatedId linter, I was expecting that the
linter was re-instantiated for each file. This assumption was incorrect!
That lead the linter to report repeated element IDs across different
files - a condition that normally happens in an app and does not
indicate a failure.

This fixes the issue and ensures we reset the state between different
files.

Closes #193