Prevent unexpected syntax from causing a crash
Sometimes people leave a `.haml` extension on a file that isn't plain
Haml. For example, you might have an ERB template that writes Haml for
a Rails generator and forget to change the extension to `.haml.erb`.
This can cause problems with linters that use the Ruby parser, since the
Haml parser cannot correctly interpret the ERB syntax of the file.

To guard against this condition, we can rescue any errors from the
parser gem and wrap them in a `Syntax` lint. This gives the user some
signal that they're doing something unexpected, but lets them lint their
entire project without haml-lint crashing on them.