Add support for an `inherits_from` directive
Borrowing from RuboCop, the idea of inheriting a configuration from
another file is useful because it allows you to have a central
configuration for your team to use across multiple projects. It's also
a necessary feature to make an `--auto-gen-config` command useful by
giving you a configuration that you can inherit from that is generated
from an existing code base.