Add `--auto-gen-config` command
Onboarding a new linter onto a project can be a pain because of
preexisting violations of the style guide. To help with this, RuboCop
has their `--auto-gen-config` commands that generates a to-do list of
violations. We should have a similar command to generate the to-do list.

Closes #119
Requires #202 for the `inherits_from` command in the readme