Support the stable version of Haml 5
Haml changed their domain model for syntax tree nodes in between
5.0.0.beta2 and 5.0.0 stable. This caused a problem with extracting the
dynamic attributes from tag nodes.

In order to maintain compatibility between Haml 4 and Haml 5, we have to
switch on the detected version of Haml and handle the dynamic attributes
differently between 4 and 5.

Haml 5 also has a subtle bug in it where the syntax tree node mutates
itself between method calls. This means that we have to memoize the
result of converting it to a literal because subsequent calls will
return a different, invalid result.