Use RuboCop severity levels to inform RuboCop lint
Currently, we set all RuboCop lint to the "warning" severity level, even
when the offense is more dire than a "warning". This changes the way
RuboCop lint is constructed so that we give a more accurate
representation of the lint to the end user.

This also adds a test to make sure we cover any new severity levels in
RuboCop and, in the case of a missing severity, defaults the lint to the
"warning" level as it used to be.

Closes #248