Make the ClassAttributeWithStaticValue more robust
Haml is extremely forgiving in its parsing, which leads to strange
errors being thrown by Haml-Lint when the parsing creates something we
weren't expecting. In this case, with a malformed set of attributes that
is still rendered by Haml, Haml-Lint was throwing an error and exiting.

By checking for the correct Haml structure as part of the inspection
predicate, we can prevent this error from happening and still yield
correct results.