Make the line span of multiline nodes more precise
Multiline nodes (particularly those using the [multiline pipe] feature
of Haml) were only reporting their beginning line in their
`#line_numbers` attribute. This meant that disabling linters via
directives was not working as expected for multiline nodes, leading to
a confusing experience for directive users.

By looking at the lines of a node, its children, its successor, and,
when necessary, the end of the document itself, we can give a more
precise end line for each node, which improves the behavior of
directives.

[multiline pipe]: http://haml.info/docs/yardoc/file.REFERENCE.html#multiline