Fix RSpec test for Rake task on Ruby 2.6
The RSpec test was breaking due to a temporary file being deleted before
all tests had run, causing errors.

This is likely due to changes in the GC implementation, but I didn't
spend too much time digging in.

To work around, modify the `config` property of the `HamlLint::RakeTask`
class to support passing either a string (like before) OR a `File`.
Allowing a `File` object allows the reference to remain ensuring that
the file is not destroyed until all tests have completed.