Fix Rake task loading
This was broken in 30eb87792 because we didn't test the case when
`config` was not explicitly specified.

I tried coming up with a test case for this but the Rake ecosystem makes
this rather odd. You can only instantiate one instance of a given Rake
task (based on my understanding) and there isn't a convenient way to
swap this out since you can't utilize a `let` in a `before(:all)`
context.

For now, I manually verified this fix works locally and will leave it at
that for now.

Fixes #300